Move models into their own files and linting
