+++ young_yacks/game/earth_system.py
@@ -1,0 +1,20 @@
+class EarthSystem:
+    name: str
+    unit: str
+    min: float
+    max: float
+    current_value: float
+
+class WaterSystem(EarthSystem):
+    unit = 'level'
+    min = 0
+    max = 100
+    # Above avegage
+    current_value = 0
+
+
+class AirSystem(EarthSystem):
+    min = 0
+    max = 100
+    unit = 'degree Celsius'
+    current_value = 0
